Added behavior test for Repo still containing entities if load fails
