+++ src/main/java/de/zalando/zmon/scheduler/ng/entities/adapters/DeployCtlProjectAdapter.java
@@ -1,7 +1,10 @@
-package de.zalando.zmon.scheduler.ng.entities;
+package de.zalando.zmon.scheduler.ng.entities.adapters;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
+import de.zalando.zmon.scheduler.ng.entities.Entity;
+import de.zalando.zmon.scheduler.ng.entities.EntityAdapter;
+import de.zalando.zmon.scheduler.ng.entities.Teams;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpEntity;
