+++ src/test/java/de/zalando/zmon/scheduler/ng/entities/EntityRepositoryTest.java
@@ -4,10 +4,11 @@ import de.zalando.zmon.scheduler.ng.config.SchedulerConfig;
 import org.junit.Test;
 import org.mockito.Mockito;
 
+import java.io.IOException;
 import java.util.List;
 
 import static java.util.Arrays.asList;
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
 /**
@@ -37,6 +38,41 @@ public class EntityRepositoryTest {
         when(registry.get("entities")).thenReturn(adapter);
 
         EntityRepository repository = new EntityRepository(registry, config);
+
+        assertEquals(1, repository.getCurrentMap().size());
+    }
+
+    @Test
+    public void TestNoChangeOnException() {
+        SchedulerConfig config = new SchedulerConfig();
+        config.setEntityBaseFilterStr("[{\"type\":\"host\"}]");
+
+        EntityAdapterRegistry registry = Mockito.mock(EntityAdapterRegistry.class);
+
+        Entity instance = new Entity("instance-1");
+        instance.addProperty("type", "instance");
+
+        Entity host = new Entity("host-1");
+        host.addProperty("type", "host");
+
+        List<Entity> entities = asList(instance, host);
+
+        EntityAdapter adapter = Mockito.mock(EntityAdapter.class);
+        when(adapter.getCollection()).thenReturn(entities).thenThrow(new RuntimeException("Loading of entities failed"));
+
+        when(registry.getSourceNames()).thenReturn(asList("entities"));
+        when(registry.get("entities")).thenReturn(adapter);
+
+        EntityRepository repository = new EntityRepository(registry, config);
+        assertEquals(1, repository.getCurrentMap().size());
+
+        try {
+            repository.fill();
+            fail("Exception not thrown");
+        }
+        catch(Exception t) {
+            assertTrue("Exception of unexpected type", t instanceof RuntimeException);
+        }
 
         assertEquals(1, repository.getCurrentMap().size());
     }
